+++ src/main/kotlin/commands/DeleteItem.kt
@@ -1,6 +1,7 @@
 package commands
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.UsageError
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
@@ -26,7 +27,11 @@ class DeleteItem(private val dataFactory: DataFactory, private val syncService: 
             if (item === null)
                 throw IdNotFoundException(itemId.toInt(), typeOf<TodoItem>())
 
-            runBlocking { syncService?.deleteItem(item.uniqueId) }
+            val response = runBlocking { syncService?.deleteItem(item.uniqueId) }
+            if (response !== null && !response.successful) {
+                throw UsageError("Deleting item failed: ${response.errorMessage}.")
+            }
+
             item.delete()
 
             terminal.println("Item deleted successfully.")
